Added splice and crop methods to image class
Also some other improvements to its interface

+++ include/wge/graphics/texture.hpp
@@ -63,7 +63,7 @@ public:
 	// Get height of texture in pixels
 	int get_height() const noexcept;
 
-	math::vec2 get_size() const noexcept;
+	math::ivec2 get_size() const noexcept;
 
 	// Set the smooth filtering. If enabled,
 	// the image will get smoothed when stretched or rotated
